backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
redis
python-dotenv
google-generativeai